Remove personal xcuserdatad directory
The xcuserdatad directory is ignored in the top level .gitignore file,
but one of its subdirectories still exist in the repository. It is not
very useful to keep such directories in a public repository since that
directory is only applicable to the user with that name. Remove the
directory from the repository.